# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Atoma API
  description: |
    REST API for Atoma - a minimalist iOS app for habit tracking, task management, and budget control.

    **Tagline**: habits, tasks, money - in one flow

    ## Authentication
    All endpoints (except auth) require a valid JWT access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Response Format
    All successful responses follow this format:
    ```json
    { "data": { ... } }
    ```

    Error responses:
    ```json
    { "error": { "code": "ERROR_CODE", "message": "..." } }
    ```
  version: 1.0.0
  contact:
    name: Atoma Support
    url: https://api.azamatbigali.online
  license:
    name: Private
servers:
  - url: https://api.azamatbigali.online/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /recurring-transactions:
    get:
      tags: [Recurring Transactions]
      summary: List Recurring Transactions
      description: Get all recurring transactions
      operationId: listRecurringTransactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of recurring transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecurringTransaction'

    post:
      tags: [Recurring Transactions]
      summary: Create Recurring Transaction
      description: Create a new recurring transaction
      operationId: createRecurringTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecurringRequest'
      responses:
        '201':
          description: Recurring transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RecurringTransaction'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /recurring-transactions/process:
    post:
      tags: [Recurring Transactions]
      summary: Process Recurring
      description: Process due recurring transactions and create actual transactions
      operationId: processRecurringTransactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Processing result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      processed:
                        type: integer
                      transactions_created:
                        type: integer
  /recurring-transactions/{id}:
    get:
      tags: [Recurring Transactions]
      summary: Get Recurring Transaction
      operationId: getRecurringTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Recurring transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RecurringTransaction'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Recurring Transactions]
      summary: Update Recurring Transaction
      operationId: updateRecurringTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecurringRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RecurringTransaction'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Recurring Transactions]
      summary: Delete Recurring Transaction
      operationId: deleteRecurringTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== GOALS ====================
...
  /savings-goal:
    get:
      tags: [Savings]
      summary: Get Savings Goal
      description: Get user's monthly savings goal with progress
      operationId: getSavingsGoal
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Savings goal with progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SavingsGoalWithProgress'

    post:
      tags: [Savings]
      summary: Set Savings Goal
      description: Create or update monthly savings goal
      operationId: setSavingsGoal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSavingsGoalRequest'
      responses:
        '200':
          description: Savings goal set
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SavingsGoalWithProgress'

    delete:
      tags: [Savings]
      summary: Delete Savings Goal
      operationId: deleteSavingsGoal
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Savings goal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean

  # ==================== AI ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
    idParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Title is required
...
  schemas:
...
    CreateRecurringRequest:
      type: object
      required: [title, amount, type, frequency]
      properties:
        title:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        frequency:
          type: string
          enum: [weekly, biweekly, monthly, quarterly, yearly]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        remaining_payments:
          type: integer
    UpdateRecurringRequest:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        frequency:
          type: string
          enum: [weekly, biweekly, monthly, quarterly, yearly]
        start_date:
          type: string
          format: date
        next_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        remaining_payments:
          type: integer
        is_active:
          type: boolean
    RecurringTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        frequency:
          type: string
          enum: [weekly, biweekly, monthly, quarterly, yearly]
        start_date:
          type: string
          format: date
        next_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        remaining_payments:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    SetSavingsGoalRequest:
      type: object
      required: [monthlyTarget]
      properties:
        monthlyTarget:
          type: number
          format: double
          minimum: 0.01
    SavingsGoalWithProgress:
      type: object
      properties:
        id:
          type: string
          format: uuid
        monthly_target:
          type: number
          format: double
        current_savings:
          type: number
          format: double
        monthly_income:
          type: number
          format: double
        monthly_expenses:
          type: number
          format: double
        progress:
          type: number
          format: double
          description: Progress 0-1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ==================== AI ====================
...
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...